some docs fixes, up FIREFOX_VERSION to 76.0 [run ci]
