+++ Makefile
@@ -22,7 +22,7 @@ CONTROL_MOCK_IMAGE_NAME := instrumentisto/medea-control-api-mock
 
 RUST_VER := 1.43
 CHROME_VERSION := 81.0
-FIREFOX_VERSION := 75.0
+FIREFOX_VERSION := 76.0
 
 crate-dir = .
 ifeq ($(crate),medea-jason)
